docs: add experimentation-service api documentation and change from int64 to int type

+++ services/experimentation-service/internal/model/event/liveEventStatistics.go
@@ -3,8 +3,8 @@ package event
 import "time"
 
 type LiveEventStatistics struct {
-	MissingRates           map[string]int64 `json:"missingRates"`
-	TotalEventsPast24Hours int64            `json:"totalEventsPast24Hours"`
-	TotalEventsPast7Days   int64            `json:"totalEventsPast7Days"`
+	MissingRates           map[string]int `json:"missingRates"`
+	TotalEventsPast24Hours int            `json:"totalEventsPast24Hours"`
+	TotalEventsPast7Days   int            `json:"totalEventsPast7Days"`
 	LastReceivedTime       time.Time        `json:"lastReceivedTime"`
 }
